Apples only roll when the player is on the same block level as them.

+++ Base/EOrange.h
@@ -11,7 +11,7 @@ public:
    EOrange(float x, float y, float w, float h, CGameplayManager* manager);
 
    bool jump;
-   float oldX;
+   double oldX;
 
    void update(float const TickTime);
    void loadMesh();
